+++ KUKAGenerator/KUKAGeneratorTest/Douglas_Peucker/Line3DTest.cpp
@@ -22,8 +22,8 @@ namespace Line3DTest
             // Arrange
 
             // these points are used to create a line
-            kuka_generator::Vector3f start(1.0f, 2.0f, 3.0f);
-            kuka_generator::Vector3f end(2.0f, 2.0f, 3.0f);
+            kuka_generator::Vector3d start(1.0f, 2.0f, 3.0f);
+            kuka_generator::Vector3d end(2.0f, 2.0f, 3.0f);
 
             // create a line from the points
             kuka_generator::CLine3D line3D(start, end);
@@ -33,13 +33,13 @@ namespace Line3DTest
             // Assert
 
             // check that the points have been stored inside the line
-            Assert::AreEqual(1.0f, line3D.p1.x);
-            Assert::AreEqual(2.0f, line3D.p1.y);
-            Assert::AreEqual(3.0f, line3D.p1.z);
+            Assert::AreEqual(1.0, line3D.p1.x);
+            Assert::AreEqual(2.0, line3D.p1.y);
+            Assert::AreEqual(3.0, line3D.p1.z);
 
-            Assert::AreEqual(2.0f, line3D.p2.x);
-            Assert::AreEqual(2.0f, line3D.p2.y);
-            Assert::AreEqual(3.0f, line3D.p2.z);
+            Assert::AreEqual(2.0, line3D.p2.x);
+            Assert::AreEqual(2.0, line3D.p2.y);
+            Assert::AreEqual(3.0, line3D.p2.z);
         }
 
         TEST_METHOD(DistanceToPointOnLineIsZeroTest)
@@ -47,8 +47,8 @@ namespace Line3DTest
             // Arrange
 
             // these points are used to create a line
-            kuka_generator::Vector3f start(1.0f, 2.0f, 3.0f);
-            kuka_generator::Vector3f end(2.0f, 2.0f, 3.0f);
+            kuka_generator::Vector3d start(1.0, 2.0, 3.0);
+            kuka_generator::Vector3d end(2.0, 2.0, 3.0);
 
             // create a line from the points
             kuka_generator::CLine3D line3D(start, end);
@@ -68,14 +68,14 @@ namespace Line3DTest
             // Arrange
 
             // these points are used to create a line
-            kuka_generator::Vector3f start(1.0f, 0.0f, 0.0f);
-            kuka_generator::Vector3f end(2.0f, 0.0f, 0.0f);
+            kuka_generator::Vector3d start(1.0, 0.0, 0.0);
+            kuka_generator::Vector3d end(2.0, 0.0, 0.0);
 
             // create a line from the points
             kuka_generator::CLine3D line3D(start, end);
 
             // create test point
-            kuka_generator::Vector3f test_point(1.0f, 1.0f, 0.0f);
+            kuka_generator::Vector3d test_point(1.0, 1.0, 0.0);
 
             // Act
